daemon: do not delete directories created by tests if tests fail
This allows for easier debugging of tests (see https://github.com/cilium/cilium/pull/9183#discussion_r323956613).

Signed-off by: Ian Vernon <ian@cilium.io>

+++ daemon/daemon_test.go
@@ -77,11 +77,7 @@ type DaemonSuite struct {
 	OnClearPolicyConsumers    func(id uint16) *sync.WaitGroup
 }
 
-func setupTestConfig() {
-	option.Config.Populate()
-	option.Config.IdentityAllocationMode = option.IdentityAllocationModeKVstore
-
-	time.Local = time.UTC
+func setupTestDirectories() {
 	tempRunDir, err := ioutil.TempDir("", "cilium-test-run")
 	if err != nil {
 		panic("TempDir() failed.")
@@ -92,15 +88,21 @@ func setupTestConfig() {
 		panic("Mkdir failed")
 	}
 
-	option.Config.DryMode = true
-	option.Config.Opts = option.NewIntOptions(&option.DaemonMutableOptionLibrary)
 	option.Config.Device = "undefined"
 	option.Config.RunDir = tempRunDir
 	option.Config.StateDir = tempRunDir
 	option.Config.AccessLog = filepath.Join(tempRunDir, "cilium-access.log")
+}
 
+func TestMain(m *testing.M) {
+	// Set up all configuration options which are global to the entire test
+	// run.
+	option.Config.Populate()
+	option.Config.IdentityAllocationMode = option.IdentityAllocationModeKVstore
+	option.Config.DryMode = true
+	option.Config.Opts = option.NewIntOptions(&option.DaemonMutableOptionLibrary)
 	// GetConfig the default labels prefix filter
-	err = labels.ParseLabelPrefixCfg(nil, "")
+	err := labels.ParseLabelPrefixCfg(nil, "")
 	if err != nil {
 		panic("ParseLabelPrefixCfg() failed")
 	}
@@ -111,9 +113,7 @@ func setupTestConfig() {
 	// state left on disk.
 	option.Config.EnableHostIPRestore = false
 
-}
-
-func TestMain(m *testing.M) {
+	time.Local = time.UTC
 	os.Exit(m.Run())
 }
 
@@ -123,7 +123,7 @@ func (epSync *dummyEpSyncher) RunK8sCiliumEndpointSync(e *endpoint.Endpoint) {}
 
 func (ds *DaemonSuite) SetUpTest(c *C) {
 
-	setupTestConfig()
+	setupTestDirectories()
 
 	ds.oldPolicyEnabled = policy.GetPolicyEnabled()
 	policy.SetPolicyEnabled(option.DefaultEnforcement)
@@ -156,7 +156,11 @@ func (ds *DaemonSuite) SetUpTest(c *C) {
 func (ds *DaemonSuite) TearDownTest(c *C) {
 	ds.d.endpointManager.RemoveAll()
 
-	os.RemoveAll(option.Config.RunDir)
+	// It's helpful to keep the directories around if a test failed; only delete
+	// them if tests succeed.
+	if !c.Failed() {
+		os.RemoveAll(option.Config.RunDir)
+	}
 
 	if ds.kvstoreInit {
 		kvstore.DeletePrefix(common.OperationalPath)
